requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.5
colorama==0.4.4
Django==3.2.7
django-environ==0.7.0
djangorestframework==3.12.4
idna==3.2
iniconfig==1.1.1
packaging==21.0
pluggy==1.0.0
psycopg2==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.6
